feat: Add "Show All Classes" and "Add Class" options to classes navigation
This commit adds the "Show All Classes" and "Add Class" options to the classes navigation in the useExperimentalLayoutLogic.tsx file. It introduces new items to the classesNavItems array, allowing users to easily navigate to the list of all classes and the page for adding a new class. This change enhances the user experience by providing convenient access to these important features.