feat: Add "Show All Evaluations" and "Add Evaluation" links to evaluations navigation
The useExperimentalLayoutLogic component in useExperimentalLayoutLogic.tsx has been updated to add "Show All Evaluations" and "Add Evaluation" links to the evaluations navigation. These links allow users to easily navigate to the respective pages for viewing all evaluations and adding a new evaluation. This enhancement improves the usability and accessibility of the application.