feat: Add deleteEntityFromVisibilityTable function to ConvexDatabase
This commit adds the deleteEntityFromVisibilityTable function to the ConvexDatabase class in order to delete an entity from the visibility table. This function takes in the options for deleting the entity and returns a Promise that resolves to either a success or a failure. This addition enhances the visibility management functionality of the application.