refactor: Update CoursUsecases to include userId in deleteSequence
This commit updates the CoursUsecases class in the cours-usecases.ts file. It adds the userId parameter to the deleteSequence method, allowing for more accurate deletion of sequences based on the user. This change enhances the functionality of the code by ensuring that only sequences belonging to the specified user are deleted.