refactor: Group competences by category
This commit adds a new function `groupCompetencesByCategory` in the `schemas.ts` file. This function takes a list of competences and returns a list of `CompetenceByCategory` objects, which contain the category name and the competences belonging to that category. This change improves the organization and grouping of competences, making it easier to work with them in the application.