refactor: Update useUpdateSequenceMetadata to redirect to the correct path
This commit updates the useUpdateSequenceMetadata function in the useUpdateSequenceMetadata.ts file to fix the redirect path when updating sequence metadata. Previously, the redirect path was missing the query parameter for the type, resulting in incorrect navigation. This has been fixed by adding the "?type=template" query parameter to the redirect path.

These changes ensure that the user is redirected to the correct page after updating sequence metadata.