refactor: Update category and competence update functions
This commit updates the updateCategory and updateCompetence functions in the category.ts and competences.ts files respectively. The functions now use the provided arguments to update the name and description of the corresponding category or competence in the database. This change enhances the functionality of the application by allowing users to easily update category and competence information.