refactor: Fix null check in ConvexDatabase class
The commit fixes a null check in the ConvexDatabase class in the convex-impl.ts file. Previously, the code checked if the result was falsy, but it should have checked if the result was null. This fix ensures that the correct condition is used for the null check, improving the reliability of the code.